+++ lib/go_http.dart
@@ -16,6 +16,7 @@ export 'src/codec/json_decoder.dart';
 export 'src/cookie/cookie_store.dart';
 export 'src/cookie/memory_cookie_store.dart';
 export 'src/decoders_stub.dart' if (dart.library.io) 'src/decoders.dart';
+export 'src/dialer.dart';
 export 'src/enrichment.dart';
 export 'src/errors.dart';
 export 'src/event_hooks.dart';